feat(dashboard): drawer carries the full review; one URL shape for both runtimes
The #472 evidence panel and decision trail landed on ReviewDetail, which no
part of the dashboard links to. grep finds zero references to
/dashboard/reviews/<id> in any component: clicking a review calls onSelect()
and opens the drawer, full stop. So the surface people use had LESS than the
one they could not reach.

That happened because #472 framed everything around ReviewDetail and cast the
drawer as the thing to borrow from. I followed the framing without asking how
a user reaches ReviewDetail — which is the question that mattered.

Fixed by extracting, not duplicating. SeverityDot, EvidencePanel and
FindingsSection were local functions inside ReviewDetail; nothing about them
was page-specific, and that accident is the only reason the drawer could not
show evidence. They now live in components/FindingEvidence.tsx and both
surfaces mount them, alongside FilterTrail which was already standalone.

This also collapses a fork rather than creating one: ReviewDetail's severity
palette was copied from the drawer during #472, and the drawer's now-dead copy
is deleted here. One palette, one findings renderer, one evidence panel.

The route stays. It cannot be deleted — it is the target of every review
comment footer, the Check Run details_url, and #468's truncation notice, which
is the only recovery path when a comment is cut. Deleting it would 404 every
comment MergeWatch has already posted. It is now a thin mount of the same
shared components.

Separately, and found by asking how this behaves self-hosted: the two runtimes
built DIFFERENT detail URLs.

  SaaS:        /dashboard/reviews/owner%2Frepo%3A42%23abc   (1 segment)
  self-hosted: /dashboard/reviews/owner%2Frepo/42#abc       (2 segments)

The route is a single [id] segment, so the self-hosted form 404'd — and the
unencoded '#' made the commit SHA a browser fragment that never reached the
server. So "View full details" has never worked on self-hosted, which also
means #468's truncation-recovery link was dead there: the one path to content
cut from an oversized comment.

Both runtimes now call buildReviewDetailUrl in core, with a test that
round-trips through the route's own parser so they cannot drift again.

Closes #486
